Changed float32 to float64 for ci_pbeam and svqbeam. Code now compiled.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/branches/SPHINX3_5_2_RCI_IRII_BRANCH@4655 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ sphinx3/src/libs3decoder/libam/fast_algo_struct.h
@@ -45,9 +45,12 @@
  * 
  * HISTORY
  * $Log$
- * Revision 1.7.4.1  2005/06/27  05:22:19  arthchan2003
- * Merged from the HEAD.
+ * Revision 1.7.4.2  2005/07/04  02:44:25  arthchan2003
+ * Changed float32 to float64 for ci_pbeam and svqbeam. Code now compiled.
  * 
+ * Revision 1.7.4.1  2005/06/27 05:22:19  arthchan2003
+ * Merged from the HEAD.
+ *
  * Revision 1.8  2005/06/22 08:00:09  arthchan2003
  * Completed all doxygen documentation on file description for libs3decoder/libutil/libs3audio and programs.
  *
@@ -273,8 +276,8 @@ extern "C" {
 			     int32 mode_dist_ds, /**< Input: (NOT USED) Whether distance-based down sampling is used */
 			     int32 isGS4GS, /**< Input: Whether Gaussian selection map is only used for Gaussian Selection. Not for down sampling */
 			     int32 isSVQ4SVQ, /**< Input: Whether SVQ is used in acoustic model computation. */
-			     float32 subvqbeam, /**< Input: Subvq beam */
-			     float32 cibeam,  /**< Input: CI phone beam */
+			     float64 subvqbeam, /**< Input: Subvq beam */
+			     float64 cibeam,  /**< Input: CI phone beam */
 			     float32 tighten_factor, /**< Input : A tightening factor used in down sampling */
 			     int32 max_cd,    /**< Input: Max CD senone to be computed */
 			     int32 n_ci_sen /**< Input: no. of ci senone, use to initialize  the ci_occ array*/
